Step clean up, sprite additions/changes, small fixes.
Added a Script called Scr_state_check

Handles all of the necessary checks including distance checks for the enemy states to be issued. This goes in tandem with the update you did, and further cleans up the step of the enemies.

Added a new sprite for the ammo, and added an animation and fade. Alarm gets set for 30 frames, enough for the fade to finish and then destroys itself. Does not affect the rest of the behavior. Also has a check if the box has been picked up or not, so even while animating, it can't be picked up multiple times.

Other various fixes and additions.